+++ src/cryptography/hazmat/backends/openssl/ec.py
@@ -3,7 +3,6 @@
 # for complete details.
 
 
-from cryptography import utils
 from cryptography.exceptions import (
     InvalidSignature,
     UnsupportedAlgorithm,
@@ -11,16 +10,10 @@ from cryptography.exceptions import (
 )
 from cryptography.hazmat.backends.openssl.utils import (
     _calculate_digest_and_algorithm,
-    _check_not_prehashed,
     _evp_pkey_derive,
-    _warn_sign_verify_deprecated,
 )
-from cryptography.hazmat.primitives import hashes, serialization
-from cryptography.hazmat.primitives.asymmetric import (
-    AsymmetricSignatureContext,
-    AsymmetricVerificationContext,
-    ec,
-)
+from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives.asymmetric import ec
 
 
 def _check_signature_algorithm(
@@ -107,49 +100,6 @@ def _ecdsa_sig_verify(backend, public_key, signature, data):
         raise InvalidSignature
 
 
-class _ECDSASignatureContext(AsymmetricSignatureContext):
-    def __init__(
-        self,
-        backend,
-        private_key: ec.EllipticCurvePrivateKey,
-        algorithm: hashes.HashAlgorithm,
-    ):
-        self._backend = backend
-        self._private_key = private_key
-        self._digest = hashes.Hash(algorithm, backend)
-
-    def update(self, data: bytes) -> None:
-        self._digest.update(data)
-
-    def finalize(self) -> bytes:
-        digest = self._digest.finalize()
-
-        return _ecdsa_sig_sign(self._backend, self._private_key, digest)
-
-
-class _ECDSAVerificationContext(AsymmetricVerificationContext):
-    def __init__(
-        self,
-        backend,
-        public_key: ec.EllipticCurvePublicKey,
-        signature: bytes,
-        algorithm: hashes.HashAlgorithm,
-    ):
-        self._backend = backend
-        self._public_key = public_key
-        self._signature = signature
-        self._digest = hashes.Hash(algorithm, backend)
-
-    def update(self, data: bytes) -> None:
-        self._digest.update(data)
-
-    def verify(self) -> None:
-        digest = self._digest.finalize()
-        _ecdsa_sig_verify(
-            self._backend, self._public_key, self._signature, digest
-        )
-
-
 class _EllipticCurvePrivateKey(ec.EllipticCurvePrivateKey):
     def __init__(self, backend, ec_key_cdata, evp_pkey):
         self._backend = backend
@@ -167,18 +117,6 @@ class _EllipticCurvePrivateKey(ec.EllipticCurvePrivateKey):
     @property
     def key_size(self) -> int:
         return self.curve.key_size
-
-    def signer(
-        self, signature_algorithm: ec.EllipticCurveSignatureAlgorithm
-    ) -> AsymmetricSignatureContext:
-        _warn_sign_verify_deprecated()
-        _check_signature_algorithm(signature_algorithm)
-        _check_not_prehashed(signature_algorithm.algorithm)
-        # This assert is to help mypy realize what type this object holds
-        assert isinstance(signature_algorithm.algorithm, hashes.HashAlgorithm)
-        return _ECDSASignatureContext(
-            self._backend, self, signature_algorithm.algorithm
-        )
 
     def exchange(
         self, algorithm: ec.ECDH, peer_public_key: ec.EllipticCurvePublicKey
@@ -272,22 +210,6 @@ class _EllipticCurvePublicKey(ec.EllipticCurvePublicKey):
     def key_size(self) -> int:
         return self.curve.key_size
 
-    def verifier(
-        self,
-        signature: bytes,
-        signature_algorithm: ec.EllipticCurveSignatureAlgorithm,
-    ) -> AsymmetricVerificationContext:
-        _warn_sign_verify_deprecated()
-        utils._check_bytes("signature", signature)
-
-        _check_signature_algorithm(signature_algorithm)
-        _check_not_prehashed(signature_algorithm.algorithm)
-        # This assert is to help mypy realize what type this object holds
-        assert isinstance(signature_algorithm.algorithm, hashes.HashAlgorithm)
-        return _ECDSAVerificationContext(
-            self._backend, self, signature, signature_algorithm.algorithm
-        )
-
     def public_numbers(self) -> ec.EllipticCurvePublicNumbers:
         get_func, group = self._backend._ec_key_determine_group_get_func(
             self._ec_key
@@ -366,6 +288,6 @@ class _EllipticCurvePublicKey(ec.EllipticCurvePublicKey):
         data, algorithm = _calculate_digest_and_algorithm(
             self._backend,
             data,
-            signature_algorithm._algorithm,  # type: ignore[attr-defined]
+            signature_algorithm.algorithm,
         )
         _ecdsa_sig_verify(self._backend, self, signature, data)